+++ CHANGELOG.markdown
@@ -4,7 +4,9 @@
 
 ### Added
 
- - Support for progress bar
+- Support for progress bar
+- `ImVec4::zero()`
+- `Into` array and tuple conversions for ImVec2 and ImVec4
 
 ### Removed